+++ python/setup.py
@@ -1,10 +1,10 @@
 from setuptools import setup
 
 setup(
-    name='pocketflow',
+    name='brainyflow',
     version='0.1.0',
-    py_modules=['pocketflow'],
+    py_modules=['brainyflow'],
     author="Victor Duarte",
     description="Minimalist LLM Framework in 65 Lines. Enable LLMs to Program Themselves.",
-    url="https://github.com/zvictor/PocketFlow",
+    url="https://github.com/zvictor/BrainyFlow",
 )
