+++ python/tests/test_flow_composition.py
@@ -1,6 +1,6 @@
 import unittest
 import asyncio
-from pocketflow import Node, Flow
+from brainyflow import Node, Flow
 
 class NumberNode(Node):
     def __init__(self, number):
