+++ typescript/tests/test_flow_basic.test.ts
@@ -1,6 +1,6 @@
 import assert from 'node:assert/strict'
 import { describe, it } from 'node:test'
-import { Flow, Node } from '../pocketflow'
+import { Flow, Node } from '../brainyflow'
 
 // Test Node implementations
 class NumberNode extends Node {
