+++ typescript/tests/test_sequential_batch_node.test.ts
@@ -1,6 +1,6 @@
 import assert from 'node:assert/strict'
 import { describe, it } from 'node:test'
-import { Node, SequentialBatchNode } from '../pocketflow'
+import { Node, SequentialBatchNode } from '../brainyflow'
 
 class TestProcessingNode extends Node {
   async exec(item: any) {
