+++ typescript/tsup.config.ts
@@ -7,6 +7,6 @@ export const tsup: Options = {
   format: ['cjs', 'esm'],
   minify: true,
   bundle: true,
-  entryPoints: ['pocketflow.ts'],
+  entryPoints: ['brainyflow.ts'],
   outDir: 'dist',
 }
